[refactor] rename routes/app/splashpage to be more intuitive
