+++ public/components/app.jsx
@@ -1,33 +1,21 @@
-import React, { Component } from 'react';
-import { Provider } from 'react-redux';
-import { createStore, combineReducers } from 'redux';
-import { Router, Route, browserHistory } from 'react-router';
-import { synchHistoryWithStore, routerReducer } from 'react-router-redux';
+import React from 'react';
+import styles from '../stylesheets/main/splash.scss';
+
+export default class Home extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
 
 
-import RentList from '../containers/rent-list';
-import reducers from '../reducers/index';
-
-const SplashPage = require('./splash_page').default;
-
-
-const store = createStore(
-          combineReducers({
-            reducers,
-            routing: routerReducer
-          })
-        );
-
-
-export default class App extends Component {
   render() {
     return (
-      <Provider store={store}>
-        <Router history={browserHistory}>
-          <Route path="/" component={SplashPage} />
-          <Route path="/listing" component={RentList} />
-        </Router>
-      </Provider>     
+      <div className='main' >
+        <form>
+          Enter your city
+          <input></input>
+        </form>
+      </div>
     );
   }
 }
