+++ public/index.jsx
@@ -1,23 +1,28 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import styles from './stylesheets/main/splash.scss';
+import { Provider } from 'react-redux';
+import { Router, browserHistory } from 'react-router';
+import { createStore, applyMiddleware } from 'redux';
 
 
-const App = require('./components/app').default;
+import routes from './components/routes';
 import reducers from './reducers';
 
 
 ReactDOM.render(
- <App />
- , document.querySelector('.container')
- );
+  <Provider store={createStore(reducers)}>
+    <Router history={browserHistory} routes={routes} />
+  </Provider>,
+  document.getElementById('.container')
+);
 
 
 if (module.hot) {
   // Whenever a new version of App.js is available
-  module.hot.accept('./components/app', () => {
+  module.hot.accept('./components/routes', () => {
     // Require the new version and render it instead
-    const NextApp = require('./components/app').default;
+    const NextApp = require('./components/routes').default;
 
     ReactDOM.render(
       <NextApp />
